version 1.3.0 just released.
SymbolLayer and HeatmapLayer were added.
The cluster were added to GeoJSONSource as an optional argument.
The color which added to Marker as an argument is seperated to optional arguments.
The number[] type of lngLat for Marker was added and there is no need to enforce type as mapboxgl.LngLatLike.
The bounds argument of the Map component is accepted not mapboxgl.LngLatBounds or mapboxgl.LngLatBoundsLike as input which means the input can also number[][] and there is no need to define type of input as standard type.
bounds and maxBounds were added to the Map Optional Arguments section.
Some problems were found while update source which is started work on it and fixed in upcoming versions.